heatmap for speed and accuracy
also took a few shots to get the right look

prompt:
- that looks great! i think you should get rid of the numbers since none of
the tests will include numbers. also can you create two heatmaps one for the
speed and one for the accuracy? since we'll have two, you should
incorporate them to the existing boxes for key speed and key accuracy. also
i think you can remove the green=fast and red=slow label, i think it will be
clear enough
- i think the box for the key stats is too short to be able to fit the
  full keyboard, i was only able to see the top row. can you fix it?
- ok that's better, i can see the full keyboard now (although i think you can
make it slightly smaller, maybe 18). however, i think the design of the
keyboard can be improved. right now it's all left-aligned so it doesn't look
like a keyboard. perhaps we can also add an ascii background to each of the
keys or space them out better? it would also be better if the letters were
spaced such that it matches the keyboard design better
- ok that looks a lot better, but now there is too much spacing between the
keys. let's keep the design but make it smaller? (maybe instead of drawing a
full ascii border, we can just use a different color for the background)