ai: add support for loading code files
prompt:
I would like to add support for passing in a file of code (for example
Ocaml code or rust code) and use that for the test data. That way users
can practice typing real code. One thing that will be important is to
not randomize the words if using this mode, as we want the text to make
sense. However, it would be good to be able to find sections of the code
to use for the test (for example parts of functions) rather than just
starting from the top every time. you can add this as an additional
option when passing in the text-source to be able to take a file.

That doesn't seem to work well with new lines. Given that code has new
lines, i think that when using the loading of a source file we should
respect the new lines in the typing test

That is really great! one problem i ran into when testing with the
test_code.ml file is that ratatype picked a part of the code that is
indented (which is fine) but i don't think it is good to require the
user to enter spaces when they are at the start of the line. instead i
think ratatype should be smart and start the cursor at the first
non-space character in each line to avoid doing that

I think there is a bug because every time I run it with the test_code.ml
file it always just starts at "let numbers = [" and it even repeats that

That looks great! Although when i run it i noticed that sometimes it
picks only a small amount of code such that I run out of code to type
within the time slot. it would be better to not run out of code to type